# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Learning Time Tracker API
  version: '1'
  description: >
    API-first backend for Learning Time Tracker. FE never talks to the
    database or file storage directly — every read/write goes through this
    API. Auth: FE signs in via Firebase client SDK (email+password or
    Google); every protected endpoint requires a Firebase ID token in
    `Authorization: Bearer <token>`, verified server-side via Firebase Admin
    SDK.
servers:
  - url: /api/v1
paths:
...
  /users/me:
    get:
      summary: Get (and lazily create) the current user's profile
      description: >
        Verifies the Firebase ID token and upserts a `users` row keyed by
        the Firebase UID if one doesn't exist yet (idempotent — safe to call
        on every sign-in, either provider). Called by the FE immediately
        after any successful sign-in.
      operationId: getCurrentUser
      tags: [users]
      security:
        - firebaseIdToken: []
      parameters:
        - name: locale
          in: query
          required: false
          description: >
            Browser-detected locale (e.g. `en`, `uk`), used only to seed the
            `locale` column when a new user row is created on first sign-in.
            Ignored for an existing user — never overwrites a persisted
            preference.
          schema:
            type: string
      responses:
        '200':
          description: The caller's user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Missing, malformed, or invalid/expired Firebase ID token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      summary: Partially update the current user's profile
      description: >
        AUTH-004 `publicName` + AUTH-008 `locale`. Every field is optional —
        OMITTING a field entirely leaves that column unchanged (partial
        update, not a full profile replace). Sending `publicName: null` (or
        an empty/whitespace-only string) explicitly resets it to `null`,
        which is a deliberate, valid action (falls back to `displayName`
        wherever this user is shown to others), not a validation error.
        Sending `locale` persists the FE header language switcher's choice
        (AUTH-008 AC4) — must be one of the registered supported codes
        (currently `en`/`uk`); the two fields update independently of one
        another in the same request.
      operationId: updateCurrentUser
      tags: [users]
      security:
        - firebaseIdToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
      responses:
        '200':
          description: The updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: >
            `publicName` longer than 100 characters
            (`errors.profile.publicNameTooLong`), or `locale` outside the
            registered supported codes (`errors.profile.localeInvalid`,
            AUTH-008 AC6) — `users.locale` is left unchanged in that case.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/me/competencies:
    get:
      summary: List the caller's own competencies
      description: >
        AUTH-005/AUTH-006/AUTH-007. Dictionary picks and free-text (custom)
        entries alike, each with its own `willingToTeach` flag. Always scoped
        to the caller (`req.firebaseUser.uid`) — there is no way to list
        another user's competencies through this API.
      operationId: listMyCompetencies
      tags: [profile]
      security:
        - firebaseIdToken: []
      responses:
        '200':
          description: The caller's competencies
          content:
            application/json:
              schema:
                type: object
                properties:
                  competencies:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserCompetency'
                required: [competencies]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Add a competency to the caller's profile
      description: >
        Exactly one of `competencyId` (a dictionary pick, AUTH-005) or
        `customLabel` (free text, AUTH-006) must be given — both or neither
        is 400 `errors.competency.invalidPayload`.


        **AUTH-005 AC3/AC5**: re-adding a `competencyId` already on the
        caller's profile is 409 `errors.competency.alreadyAdded` (no
        upsert/idempotency here, unlike board_members/task_shares) — an
        unknown or inactive `competencyId` is 400
        (`errors.competency.notFound` / `errors.competency.inactive`
        respectively). A blank/whitespace-only `customLabel` is 400
        `errors.competency.customLabelRequired`; longer than 100 characters is
        `errors.competency.customLabelTooLong`. Duplicate `customLabel`
        values ARE allowed (AUTH-006 AC7) — free text is never deduplicated.
      operationId: addMyCompetency
      tags: [profile]
      security:
        - firebaseIdToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCompetencyCreateRequest'
      responses:
        '201':
          description: The created competency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCompetency'
        '400':
          description: >
            Invalid payload shape, unknown/inactive competencyId, or invalid
            customLabel — see the messageKey table above.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: This `competencyId` is already on the caller's profile (`errors.competency.alreadyAdded`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/me/competencies/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      summary: Toggle a competency's "willing to teach" flag
      description: >
        AUTH-007. Works for both dictionary and custom competencies (AC4) —
        the flag doesn't depend on origin.


        **Anti-enumeration (AC3)**: an id belonging to another user's
        `user_competencies` row resolves to the SAME 404
        `errors.competency.notFound` as a genuinely missing id — never 403.
        Same convention already applied to `time_entries` (US-011 AC7) and
        to sign-in (AUTH-002 AC2).
      operationId: updateMyCompetencyWillingToTeach
      tags: [profile]
      security:
        - firebaseIdToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCompetencyWillingToTeachRequest'
      responses:
        '200':
          description: The updated competency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCompetency'
        '400':
          description: '`willingToTeach` missing or not a boolean (`errors.competency.invalidPayload`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No such competency on the caller's own profile (see anti-enumeration note above)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Remove a competency from the caller's profile
      description: >
        AUTH-005 AC4 / AUTH-006 AC6 — a single endpoint for both dictionary
        and custom competencies. Removing a competency that has
        `willingToTeach = true` simply removes the row; the flag has no
        existence independent of it. Same anti-enumeration contract as PATCH
        above.
      operationId: removeMyCompetency
      tags: [profile]
      security:
        - firebaseIdToken: []
      responses:
        '204':
          description: Removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No such competency on the caller's own profile (see anti-enumeration note above)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/search:
    get:
      summary: Find public profiles willing to teach a competency
      description: >
        US-025 AC1/AC4. Every `users` row with a `user_competencies` row for
        `competencyId` that has `willing_to_teach = true` — literally "ready
        to teach it", not merely "has it listed" (AC1). Each result also
        carries that user's FULL list of `willing_to_teach = true`
        competencies, not just the one searched for (AC4 — useful context).
        Never exposes email, boards, or time data (AC4). No pagination in
        MVP.


        A missing or malformed `competencyId` resolves to an empty result
        list rather than a 400 — the FE is expected not to call this without
        a selection (AC2), but there's no AC-defined error path for a
        malformed one either, and "empty" is indistinguishable in effect
        from AC3's "no matches" empty state.
      operationId: searchUsersByCompetency
      tags: [users]
      security:
        - firebaseIdToken: []
      parameters:
        - name: competencyId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Matching public profiles (possibly empty)
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicUserSearchResult'
                required: [users]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Firebase UID (not a UUID — `users.id` is the raw Firebase UID)
        schema:
          type: string
    get:
      summary: View another user's public profile
      description: >
        US-026 AC1/AC2. Available for ANY valid `userId`, including the
        caller's own — deliberately NOT gated by shared board/task
        membership (AC1's "доступний для БУДЬ-ЯКОГО валідного userId"), the
        one intentionally wide-open cross-user read in this API. Returns the
        FULL competency list (unlike `/users/search`, which only returns
        willing-to-teach ones) with a per-entry `willingToTeach` flag so the
        FE can render the "Готовий(а) викладати" badge. Never exposes email
        or any other private field.
      operationId: getPublicUserProfile
      tags: [users]
      security:
        - firebaseIdToken: []
      responses:
        '200':
          description: The public profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No user with this id (`errors.user.notFound`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  responses:
    Unauthorized:
      description: Missing, malformed, or invalid/expired Firebase ID token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  securitySchemes:
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token, verified via Firebase Admin SDK.
  schemas:
    UserProfile:
      type: object
      properties:
        id:
          type: string
          description: Firebase UID
        email:
          type: string
          format: email
        displayName:
          type: string
          description: System-generated, never user-editable (see the users table comment in CLAUDE.md "Дані").
        publicName:
          type: string
          nullable: true
          description: >
            AUTH-004 — user-editable name shown to other users in shared
            contexts (board_members/task_shares listings) INSTEAD OF
            displayName, when set. Null falls back to displayName wherever
            this user is shown to someone else.
        locale:
          type: string
          example: en
        createdAt:
          type: string
          format: date-time
      required: [id, email, displayName, publicName, locale, createdAt]
    UserProfileUpdateRequest:
      type: object
      description: >
        Partial update — every field is optional; an OMITTED field is left
        unchanged (AUTH-004 AC8, AUTH-008 AC4). Recognizes `publicName` and
        `locale`, updated independently of one another.
      properties:
        publicName:
          type: string
          nullable: true
          maxLength: 100
          description: >
            An explicit `null` (or an empty/whitespace-only string) resets to
            the displayName fallback (AC3) — not a validation error. Over 100
            characters (after trim) is 400 `errors.profile.publicNameTooLong`.
        locale:
          type: string
          description: >
            AUTH-008 — the FE header language switcher's active locale.
            Must be one of the registered supported codes (currently
            `en`/`uk`); any other value is 400
            `errors.profile.localeInvalid` and leaves `users.locale`
            unchanged.
          example: uk
...
    UserCompetency:
      type: object
      description: >
        One of the caller's own competencies — either a dictionary pick
        (`isCustom = false`, `competencyId`/`competencySlug` set) or a
        free-text entry (`isCustom = true`, `customLabel` set). Exactly one
        of `competencySlug`/`customLabel` is non-null on any given row.
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          description: Firebase UID — always the caller for any row this API ever returns.
        competencyId:
          type: string
          format: uuid
          nullable: true
        competencySlug:
          type: string
          nullable: true
          description: Set only when isCustom is false — render via `competency.<slug>`, same as Competency.slug above.
        isCustom:
          type: boolean
        customLabel:
          type: string
          nullable: true
          description: Set only when isCustom is true. User-generated text — rendered as-is, never run through the locale dictionary.
        willingToTeach:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, competencyId, competencySlug, isCustom, customLabel, willingToTeach, createdAt, updatedAt]
    UserCompetencyCreateRequest:
      type: object
      description: Exactly one of `competencyId`/`customLabel` — see POST /users/me/competencies for the full validation rules.
      properties:
        competencyId:
          type: string
          format: uuid
        customLabel:
          type: string
          maxLength: 100
    UserCompetencyWillingToTeachRequest:
      type: object
      properties:
        willingToTeach:
          type: boolean
      required: [willingToTeach]
...
    PublicUserCompetency:
      type: object
      description: >
        One entry in a public profile's competency list (US-025/US-026) —
        deliberately a narrower shape than the caller's-own-profile
        `UserCompetency` schema: no `id`/`userId`/timestamps, since those
        have no meaning to a viewer who isn't the owner.
      properties:
        competencyId:
          type: string
          format: uuid
          nullable: true
          description: 'Null for a custom (`isCustom: true`) entry.'
        competencySlug:
          type: string
          nullable: true
          description: Locale key suffix (`competency.<slug>`); null for a custom entry.
        isCustom:
          type: boolean
        customLabel:
          type: string
          nullable: true
        willingToTeach:
          type: boolean
      required: [competencyId, competencySlug, isCustom, customLabel, willingToTeach]
    PublicUserSearchResult:
      type: object
      description: >
        One result row from `GET /users/search` (US-025 AC4). Only carries
        the `willing_to_teach = true` competencies (every entry here has
        `willingToTeach: true`) — contrast with `PublicUserProfile` below,
        which carries the FULL list. Never email/boards/time.
      properties:
        id:
          type: string
          description: Firebase UID
        name:
          type: string
          description: publicName if set, else displayName (AUTH-004 AC5/AC6 pattern).
        competencies:
          type: array
          items:
            $ref: '#/components/schemas/PublicUserCompetency'
      required: [id, name, competencies]
    PublicUserProfile:
      type: object
      description: >
        `GET /users/{id}` (US-026 AC1). The FULL `user_competencies` list
        (not just willing-to-teach ones, unlike PublicUserSearchResult) —
        each entry's own `willingToTeach` flag is what the FE renders the
        badge from. Never email/boards/time.
      properties:
        id:
          type: string
          description: Firebase UID
        name:
          type: string
          description: publicName if set, else displayName (AUTH-004 AC5/AC6 pattern).
        competencies:
          type: array
          items:
            $ref: '#/components/schemas/PublicUserCompetency'
      required: [id, name, competencies]
...
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: AUTH_TOKEN_INVALID
            messageKey:
              type: string
              description: Localization key the FE looks up in its own locale dictionary.
              example: errors.auth.tokenInvalid
            message:
              type: string
              description: English fallback for logs/devtools — not shown to end users.
          required: [code, messageKey]
      required: [error]
security:
  - firebaseIdToken: []

# ---------------------------------------------------------------------------
# WebSocket channel (US-029) — documented here as plain text, NOT as an
# OpenAPI `paths` entry, since OpenAPI 3.0 has no syntax for a WS
# upgrade/frame protocol. This is the one deliberate, explicitly-documented
# departure from "every endpoint is a REST path in this file" (CLAUDE.md
# "Архітектура": the FE ↔ BE contract was pure REST before US-029).
#
# Placement decision (US-029 AC6/AC7): the WS server is attached to the
# SAME http.Server/port (4000) the Express app already listens on — see
# backend/src/ws/server.js's header comment for the full rationale and
# trade-offs. `docker-compose.yml` is therefore UNCHANGED by this pass;
# there is no new service to add.
#
# Endpoint:      ws(s)://<same host:port as the REST API>/ws?token=<Firebase ID token>
#
# Authentication (US-029 AC2/AC3): the SAME Firebase ID token used for
# `Authorization: Bearer` on REST calls, passed as the `token` query
# parameter (the browser's native WebSocket API can't set custom headers on
# the upgrade request). Verified via the same Firebase Admin SDK
# `verifyIdToken()` the REST `requireAuth` middleware uses — not a separate
# auth mechanism. A missing or invalid/expired token gets a structured
# `error` frame (`ws.error.unauthorized`) followed immediately by the
# server closing the connection with close code 4401 — no subscription or
# message is ever accepted first.
#
# Delivery guarantee (US-027 AC8/AC9, US-029 AC5/AC9): REST is unconditionally
# the source of truth. Every message is written to Postgres by a REST POST
# BEFORE any WS event is broadcast; a missing/dropped/never-subscribed WS
# connection never loses a message — the full history is always available
# via `GET /dm-threads/{id}/messages` / `GET /competencies/{id}/chat/messages`.
# WS only accelerates "live" delivery to connections that happen to be
# listening at that instant.
#
# --- Client -> Server frames ---
#
#   {"type": "subscribe", "channel": "dmThread", "threadId": "<uuid>"}
#     Authorized via the SAME service-layer check
#     (`requireDmThreadAccess`) `GET /dm-threads/{id}/messages` uses — only
...